Initial commit: just the README
So I sort of maybe kinda wrote the README exactly how you're not
supposed to write code: the entire thing in one commit over many days.

It was an accident! Really! At first I just wanted a README with a basic
description for the initial commit, but my description was all about the
differences from git-subtree and assumed you knew what even the
high-level point of git-subtree or git-subhistory were, so then I went
back to write an intro so you'd know what this was about even if you'd
never even heard of git-subtree, but then I had to finish my comparison
with git-subtree 'cause I can't just have half an explanation of that
hanging out at the end, but then I had to explain in detail how
`merge -s subtree` works and why it's bad that `git-subtree merge` uses
it, but all the explanations sucked so I needed an example, and so here
we are.