Add skeleton test script; .gitignore its test repo
Not based on anything, especially not git-subtree/test.sh. In my humble
opinion, tools like tests are only as useful as what you read, and
output that you habitually skim is worse than useless, it's clutter that
can hide actual important information, so a "just the highlights" view
is crucial.