Move Descriptions above Example in README
The example is dumb and bad; it's unclear to a reader who doesn't know
what to expect, and is unhelpful to a reader who knows what to expect
and will be worried about corner cases, and because it lacks corner
cases is a worthless test case.

But for now move it below the descriptions.