Move parsing <subproj-path> back into subcommands
Whoops, just realized that unlike 'split', 'merge' and 'assimilate'
should work even if path/to/sub/ doesn't exist at all, like when
importing in somebody else's library as a subproject, or when
transitioning from a submodule to subhistory.

(This just preps for that, that still needs specialized code in
subhistory_assimilate().)

Incredibly, there's appears to be no universal way in shell to normalize
away .'s, ..'s, //'s, etc. in paths, nor is this gaping hole in basic
functionality plugged by any git utilities, cd+pwd appears to be the
best way to do that. (There's a C function, realpath(), which there are
binary wrappers for, and many Linux platforms have readlink, but no
universal builtin that I can find.) This is problematic because
'assimilate', like 'split' shouldn't change the working tree in any way
at all, but needs the path to be normalized (git-read-tree is picky).

It's not a huge deal because the real use of 'assimilate' is 'merge',
which has to create the subdirectory to merge the subproject into
anyway, but it's frustrating because it's not the "right thing" for
'assimilate', and is entirely due to an outrageous limitation of shell
scripting.