Fix assimilating subprojs that weren't from split
like when importing in somebody else's library as a subproject, or when
transitioning from a submodule to subhistory.

Recommend -w to view diff.

Basically, check if "path/to/sub/" has history at all yet, if so do the
original thing, if not just rewrite every commit with its tree moved
under "path/to/sub/" (like the last example of `git help filter-branch`,
but simpler and in a cross-platform way: only GNU sed does what you want
to \t like used in that example, not BSD sed. I should submit a
documentation patch for that).

Also, subhistory_merge() makes sure "path/to/sub/" already exists for
get_path_to_sub() in subhistory_assimilate().