Assim'd merge commits w/same parent Main tree keep it
Instead of, when assimilating any merge commit at all, using HEAD as
the Main tree, first check if every parent of the merge commit has the
same Main tree, and if so just use that as the Main tree, otherwise
default to HEAD as the Main tree.

Recommend --word-diff to view diff.

Change index_filter comment footnote marker from just [Footnote] to
[FN1], [FN2], so can add footnote about the new code to check if every
parent of a merge commit has the same Main tree.

Also fix silencing `git rev-parse --verify -q $GIT_COMMIT^2`, for which
apparently the -q keeps it from erroring when it fails to verify but it
still prints the SHA to stdout when it succeeds. (First test case
assimilating a merge commit.)