Use indent rather than ``` fences for code blocks in README.md
Code fences ``` is mainly useful to specify syntax highlighting in
GitHub-Flavored Markdown. The 4-space indent syntax is compatible with
the original markdown, and is also more readable in plaintext IMO.

Also, get rid of the 2-space indent in the shell examples (leftover from
before we switched to markdown from plaintext formatting).